React-Django-Mysql/DjangoReact_project/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
cffi==1.15.0
configparser==5.2.0
cryptography==36.0.1
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
pycparser==2.21
PyMySQL==1.0.2
pytz==2021.3
sqlparse==0.4.2
